--- src/main/java/prgrms/marco/be02marbox/domain/Region.java
@@ -1,5 +1,0 @@
-package prgrms.marco.be02marbox.domain;
-
-public enum Region {
-	SEOUL, BUSAN;
-}
