+++ src/main/java/prgrms/marco/be02marbox/domain/movie/Genre.java
@@ -1,4 +1,4 @@
-package prgrms.marco.be02marbox.domain;
+package prgrms.marco.be02marbox.domain.movie;
 
 public enum Genre {
 	ACTION, ROMANCE, ANIMATION
