+++ src/main/java/prgrms/marco/be02marbox/domain/movie/LimitAge.java
@@ -1,0 +1,5 @@
+package prgrms.marco.be02marbox.domain.movie;
+
+public enum LimitAge {
+	ADULT, CHILD
+}
