+++ src/main/java/prgrms/marco/be02marbox/domain/reservation/ReservedSeat.java
@@ -1,14 +1,14 @@
-package prgrms.marco.be02marbox.domain;
+package prgrms.marco.be02marbox.domain.reservation;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
+import prgrms.marco.be02marbox.domain.theater.Seat;
 
 @Entity
 @Table(name = "reserved_seat")
