+++ src/main/java/prgrms/marco/be02marbox/domain/theater/Region.java
@@ -1,0 +1,5 @@
+package prgrms.marco.be02marbox.domain.theater;
+
+public enum Region {
+	SEOUL, BUSAN;
+}
