+++ src/main/java/prgrms/marco/be02marbox/domain/theater/Theater.java
@@ -1,4 +1,4 @@
-package prgrms.marco.be02marbox.domain;
+package prgrms.marco.be02marbox.domain.theater;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -10,28 +10,18 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "movie")
-public class Movie {
+@Table(name = "theater")
+public class Theater {
 
 	@Id
 	@Column(name = "id")
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
 
+	@Column(name = "region")
+	@Enumerated(value = EnumType.STRING)
+	private Region region;
+
 	@Column(name = "name")
 	private String name;
-
-	@Column(name = "limit_age")
-	@Enumerated(value = EnumType.STRING)
-	private LimitAge limitAge;
-
-	@Column(name = "genre")
-	@Enumerated(value = EnumType.STRING)
-	private Genre genre;
-
-	@Column(name = "running_time")
-	private Long runningTime;
-
-	@Column(name = "poster_img_location")
-	private String posterImgLocation;
 }
