+++ src/main/java/prgrms/marco/be02marbox/domain/theater/repository/TheaterRepository.java
@@ -1,0 +1,8 @@
+package prgrms.marco.be02marbox.domain.theater.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import prgrms.marco.be02marbox.domain.theater.Theater;
+
+public interface TheaterRepository extends JpaRepository<Theater, Long> {
+}
