+++ src/main/java/prgrms/marco/be02marbox/domain/theater/repository/TheaterRoomRepository.java
@@ -1,0 +1,8 @@
+package prgrms.marco.be02marbox.domain.theater.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import prgrms.marco.be02marbox.domain.theater.TheaterRoom;
+
+public interface TheaterRoomRepository extends JpaRepository<TheaterRoom, Long> {
+}
