+++ src/main/java/prgrms/marco/be02marbox/domain/user/Role.java
@@ -1,4 +1,4 @@
-package prgrms.marco.be02marbox.domain;
+package prgrms.marco.be02marbox.domain.user;
 
 public enum Role {
 	CUSTOMER, MANAGER, ADMIN
