+++ src/main/java/prgrms/marco/be02marbox/domain/user/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package prgrms.marco.be02marbox.domain.user.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import prgrms.marco.be02marbox.domain.user.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
